+++ src/main/java/org/oagi/srt/gateway/http/api/cc_management/data/CcAccRequest.java
@@ -2,7 +2,9 @@ package org.oagi.srt.gateway.http.api.cc_management.data;
 
 import lombok.Data;
 
+import java.math.BigInteger;
+
 @Data
 public class CcAccRequest {
-    private Long basedAccManifestId;
+    private BigInteger basedAccManifestId;
 }
